+++ Dolphin scripts/Entrance Randomizer/lib/types_.py
