+++ Dolphin scripts/Entrance Randomizer/lib/utils.py
@@ -1,0 +1,44 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from dolphin import gui  # pyright: ignore[reportMissingModuleSource]
+from lib.constants import *  # noqa: F403
+from lib.constants import __version__
+from lib.types_ import SeedType
+
+DRAW_TEXT_STEP = 24
+DRAW_TEXT_OFFSET_X = 272
+_draw_text_index = 0
+"""Count how many times draw_text has been called this frame"""
+
+
+def reset_draw_text_index():
+    global _draw_text_index
+    _draw_text_index = 0
+
+
+def draw_text(text: str):
+    global _draw_text_index
+    gui.draw_text(
+        (DRAW_TEXT_OFFSET_X, DRAW_TEXT_STEP / 2 + DRAW_TEXT_STEP * _draw_text_index),
+        0xFF00FFFF,
+        text,
+    )
+    _draw_text_index += 1
+
+
+def dump_spoiler_logs(starting_area_name: str, transitions_map: dict[int, dict[int, int]], seed_string: SeedType):
+    spoiler_logs = f"Starting area: {starting_area_name}\n"
+    for from_, to_old_and_new in transitions_map.items():
+        for to_old, to_new in to_old_and_new.items():
+            spoiler_logs += f"From: {TRANSITION_INFOS_DICT[from_].name}, " + \
+                f"To: {TRANSITION_INFOS_DICT[to_old].name}. " + \
+                f"Redirecting to: {TRANSITION_INFOS_DICT[to_new].name}\n"
+
+    dolphin_path = Path().absolute()
+    print("path in utils", dolphin_path)
+    spoiler_logs_file = dolphin_path / "User" / "Logs" / f"SPOILER_LOGS_v{__version__}_{seed_string}.txt"
+    Path.mkdir(spoiler_logs_file.parent, exist_ok=True)
+    Path.write_text(spoiler_logs_file, spoiler_logs)
+    print("Spoiler logs written to", spoiler_logs_file)
